+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 - DefaultConditionModel.DefaultSetCondition.equals() and hashCode() now throw UnsupportedOperationException, in case equalTo() is confused with equals().
 - DefaultConditionModel bug fixed, adding wildards no longer breaks case-insensitivity.
 - DefaultConditionModel bug fixed, in now works when case-insensitive.
+- DefaultConditionModel bug fixed, now adjusts Temporal precision to match the dateTimePattern.
 ### is.codion.swing.common.model
 - DefaultListSelection.SelectedIndexes and DefaultItems.optional() now return an empty Optional in case of empty selection.
 - DefaultListSelection.SelectedIndex no longer non-null, returns null instead of -1 in case of empty selection.